Moved Compressor to store package
