+++ backend/src/main/java/com/empire/store/Compressor.java
@@ -1,4 +1,4 @@
-package com.empire;
+package com.empire.store;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
